Do not forward pre-flight requests to handlers
This request is only used by the client to determine if it is allowed
to send a cross-origin request.  In the case of OPTIONS, then the
client will send a second OPTIONS request that should then be served
by the handler.

In the meantime, the preflight tests for a default policy have been
abandoned because it is unclear whether or not to pass a request to
the handler when the origin does not match.